refactor: split get test code
